requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0 
pydantic>=2.0.0
pint>=0.10.0
fooddatacentral>=1.0.0